snippet worked on arm64 windows.

+++ src/platform.h
@@ -66,7 +66,7 @@ constexpr const std::string_view env_home = "userprofile";
 constexpr const std::string_view path_split = ";";
 constexpr const std::string_view ndk_build = "ndk-build.bat";
 
-static inline uint32_t mem_page_size_impl() {
+static inline uint64_t mem_page_size_impl() {
   SYSTEM_INFO info;
   ::GetSystemInfo(&info);
   return info.dwPageSize;
